+++ CHANGELOG.md
@@ -32,3 +32,4 @@
 * Fixed a bug where the path to the source files was not calculated correctly if `MultipleCopies` was set on `Copy-AxiumFiles`.
 
 # Version 1.4.0
+* Added ability to check if a copy of axiUm is installed in a given folder. Fixes issue #28.